+++ lib/src/install_solvro/add_linter.dart
@@ -2,9 +2,10 @@ import "dart:io";
 
 import "package:mason_logger/mason_logger.dart";
 
-Future<void> addLinter(String yamlName, Logger logger) async {
-  await File(
-    "analysis_options.yaml",
-  ).writeAsString("include: package:solvro_config/$yamlName.yaml");
-  logger.info(green.wrap("Overriden analysis_options.yaml"));
+import "utils/add_if_not_exist.dart";
+
+void addLinter(String yamlName, Logger logger) {
+  final template = "include: package:solvro_config/$yamlName.yaml";
+  final file = File("analysis_options.yaml");
+  addIfNotExist(logger, file, template);
 }