+++ lib/src/install_solvro/ci/add_ci_if_not_exist.dart
@@ -1,0 +1,10 @@
+import "dart:io";
+
+import "package:mason_logger/mason_logger.dart";
+
+import "../utils/add_if_not_exist.dart";
+
+void addCI(Logger logger, String workflowName, String workflowContent) {
+  final workflowFile = File(".github/workflows/$workflowName.yml");
+  return addIfNotExist(logger, workflowFile, workflowContent);
+}